use case 2 and 3 except borrow and review fn

+++ src/main/java/com/Product/BookStore/repository/BookRepository.java
@@ -3,7 +3,6 @@ package com.Product.BookStore.repository;
 import com.Product.BookStore.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book,Long> {
-
+public interface BookRepository extends JpaRepository<Book,Integer> {
 
 }
